Split thing creation into specific classes.
